+++ custom_components/dashboardmodern/frontend/legacy/build-info.js
@@ -13,4 +13,4 @@ import "../src/sections/beta25-real-device-fixes-section.js";
 import "../src/sections/beta25-compatibility-section.js";
 import "../src/sections/beta26-real-device-stability-section.js";
 import "../src/sections/segnalazioni-section.js";
-export const BUILD_INFO = Object.freeze({"generated":true,"integrationVersion":"1.4.5-beta.10","dashboardVersion":"1.4.5-beta.10","moduleVersion":14,"schemaVersion":4,"date":"2026-09-02T11:13:09+00:00","commit":"1b158eff8d0b37bb011bd5f86476835a53e9f373","assetHash":"e6ff445e4565d6cc"});
+export const BUILD_INFO = Object.freeze({"generated":true,"integrationVersion":"1.4.5-beta.10","dashboardVersion":"1.4.5-beta.10","moduleVersion":14,"schemaVersion":4,"date":"2026-09-02T11:19:52+00:00","commit":"0a61c110fa22cb684f1c7b7073e65977616c70ec","assetHash":"85e52e483888e24a"});